Updated Triangulo.cs

+++ src/Application/Triangulo.cs
@@ -2,14 +2,14 @@ namespace Application;
 
 public class Triangulo
 {
-    public static double Perimetro(double lado1, double lado2, double lado3)
+    public static float Perimetro(float lado1, float lado2, float lado3)
     {
-        double perimetro = lado1 + lado2 + lado3;
+        float perimetro = lado1 + lado2 + lado3;
         return perimetro;
     }
-    public static double Area(double _base, double altura)
+    public static float Area(float _base, float altura)
     {
-        double area = ((_base * altura) / 2);
+        float area = ((_base * altura) / 2);
         return area;
     }
 }
